+++ app/app.csproj
@@ -11,7 +11,6 @@
   <PropertyGroup Condition=" '$(RunConfiguration)' == 'EscolaService' " />
 
   <ItemGroup>
-    <PackageReference Include="CsvHelper" Version="30.0.1" />
     <PackageReference Include="Enums.NET" Version="4.0.1" />
     <PackageReference Include="Hangfire.AspNetCore" Version="1.8.6" />
     <PackageReference Include="Hangfire.Core" Version="1.8.6" />
