+++ test/Fixtures/Base.cs
@@ -1,47 +1,65 @@
 ﻿using api;
+using app.Controllers;
+using app.Entidades;
+using app.Repositorios;
+using app.Repositorios.Interfaces;
 using app.Services;
+using app.Services.Interfaces;
 using auth;
+using Hangfire;
+using Microsoft.EntityFrameworkCore;
+using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.DependencyInjection;
+using service.Interfaces;
 using System.Collections.Generic;
 using System.IdentityModel.Tokens.Jwt;
 using System.Linq;
 using System.Security.Claims;
+using System.Threading.Tasks;
 using Xunit.Abstractions;
+using Xunit.Microsoft.DependencyInjection;
 using Xunit.Microsoft.DependencyInjection.Abstracts;
 
 namespace test.Fixtures
 {
-    public class AuthTest : TestBed<Base>
+    public class Base : TestBedFixture
     {
-        ClaimsPrincipal Usuario;
-        AuthService authService;
-
-        public AuthTest(ITestOutputHelper testOutputHelper, Base fixture) : base(testOutputHelper, fixture)
+        protected override void AddServices(IServiceCollection services, IConfiguration? configuration)
         {
-            authService = fixture.GetService<AuthService>(testOutputHelper)!;
+            // Para evitar a colisão durante a testagem paralela, o nome deve ser diferente para cada classe de teste
+            var databaseName = "DbInMemory" + Random.Shared.Next().ToString();
+            services.AddDbContext<AppDbContext>(o => o.UseInMemoryDatabase(databaseName));
+
+            // Repositorios
+            services.AddScoped<IEscolaRepositorio, EscolaRepositorio>();
+            services.AddScoped<IMunicipioRepositorio, MunicipioRepositorio>();
+            services.AddScoped<IRanqueRepositorio, RanqueRepositorio>();
+            services.AddScoped<ISuperintendenciaRepositorio, SuperIntendenciaRepositorio>();
+
+            // Services
+            services.AddScoped<IEscolaService, EscolaService>();
+            services.AddScoped<IMunicipioService, MunicipioService>();
+            services.AddScoped<ISolicitacaoAcaoService, SolicitacaoAcaoService>();
+            services.AddScoped<IRanqueService, RanqueService>();
+            services.AddScoped<IUpsService, UpsServiceMock>();
+            services.AddScoped<IBackgroundJobClient, BackgroundJobClientFake>();
+            services.AddSingleton<ModelConverter>();
+            services.AddScoped<ISuperintendenciaService, SuperintendenciaService>();
+
+            // Controllers
+            services.AddScoped<DominioController>();
+            services.AddScoped<EscolaController>();
+            services.AddScoped<RanqueController>();
+            services.AddScoped<SuperintendenciasController>();
+
+            services.AddAuth(configuration);
         }
 
-        public (string Token, ClaimsPrincipal Usuario) AutenticarUsuario(AppController controller, AuthUserModel<Permissao>? usuario = null, List<Permissao>? permissoes = null)
+        protected override ValueTask DisposeAsyncCore() => new();
+
+        protected override IEnumerable<TestAppSettings> GetTestAppSettings()
         {
-            if (usuario == null) {
-                usuario = new AuthUserModel<Permissao>
-                {
-                    Id = 1,
-                    Name = "test",
-                    Permissions = Enum.GetValues<Permissao>().ToList(),
-                };
-            }
-            if (permissoes != null)
-            {
-                usuario.Permissions = permissoes;
-            }
-            
-            var (token, _) = authService.GenerateToken(usuario);
-
-            var jwt = new JwtSecurityTokenHandler().ReadJwtToken(token);
-
-            Usuario = new ClaimsPrincipal(new ClaimsIdentity(jwt.Claims));
-            controller.AppUsuario = Usuario;
-            return (token, Usuario);
+            yield return new() { Filename = "appsettings.Test.json", IsOptional = false };
         }
     }
 }
